requirements.txt
transformers==4.30.2
torch>=2.0
faiss-cpu
requests
python-dotenv
pdfplumber
easyocr
pillow